requirements.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
python-dotenv
Requests
GitHub-Flask
Flask-Migrate